Fe fold

+++ be/test/vec/function/function_time_test.cpp
@@ -179,7 +179,7 @@ TEST(VTimestampFunctionsTest, from_unix_test) {
     std::string func_name = "from_unixtime";
     TimezoneUtils::load_timezone_names();
 
-    InputTypeSet input_types = {TypeIndex::Int32};
+    InputTypeSet input_types = {TypeIndex::Int64};
 
     DataSet data_set = {{{1565080737}, std::string("2019-08-06 16:38:57")}, {{-123}, Null()}};
